+++ tests/test_gg.py
@@ -2208,6 +2208,65 @@ def test_oldrperp_local():
     np.testing.assert_allclose(corr2_output['xip_im'], gg.xip_im, rtol=1.e-3)
 
 
+def test_varxi():
+    # Test that varxi is correct (or close) based on actual variance of 100 runs.
+
+    # Same gamma pattern as in test_gg().  Although the signal doesn't actually matter at all here.
+    gamma0 = 0.05
+    r0 = 10.
+    L = 50.*r0
+    np.random.seed(8675309)
+
+    # Note: to get a good estimate of var(xi), you need a lot of runs.  The number of
+    # runs matters much more than the number of galaxies for getting this to pass.
+    if __name__ == '__main__':
+        ngal = 1000
+        nruns = 50000
+        tol_factor = 1
+    else:
+        ngal = 100
+        nruns = 5000
+        tol_factor = 5
+
+    all_ggs = []
+    for run in range(nruns):
+        # In addition to the shape noise below, there is shot noise from the random x,y positions.
+        x = (np.random.random_sample(ngal)-0.5) * L
+        y = (np.random.random_sample(ngal)-0.5) * L
+        r2 = (x**2 + y**2)/r0**2
+        g1 = -gamma0 * np.exp(-r2/2.) * (x**2-y**2)/r0**2
+        g2 = -gamma0 * np.exp(-r2/2.) * (2.*x*y)/r0**2
+        # This time, add some shape noise (different each run).
+        g1 += np.random.normal(0, 0.3, size=ngal)
+        g2 += np.random.normal(0, 0.3, size=ngal)
+
+        cat = treecorr.Catalog(x=x, y=y, g1=g1, g2=g2, x_units='arcmin', y_units='arcmin')
+        gg = treecorr.GGCorrelation(bin_size=0.1, min_sep=10., max_sep=100., sep_units='arcmin',
+                                    verbose=1)
+        gg.process(cat)
+        all_ggs.append(gg)
+        print('.', end='', flush=True)
+    print()
+
+    mean_xip = np.mean([gg.xip for gg in all_ggs], axis=0)
+    var_xip = np.var([gg.xip for gg in all_ggs], axis=0)
+    mean_xim = np.mean([gg.xim for gg in all_ggs], axis=0)
+    var_xim = np.var([gg.xim for gg in all_ggs], axis=0)
+    mean_varxi = np.mean([gg.varxi for gg in all_ggs], axis=0)
+
+    print('mean_xip = ',mean_xip)
+    print('mean_xim = ',mean_xim)
+    print('mean_varxi = ',mean_varxi)
+    print('var_xip = ',var_xip)
+    print('ratio = ',var_xip / mean_varxi)
+    print('var_xim = ',var_xim)
+    print('ratio = ',var_xim / mean_varxi)
+    print('max relerr for xip = ',np.max(np.abs((var_xip - mean_varxi)/var_xip)))
+    print('max relerr for xim = ',np.max(np.abs((var_xim - mean_varxi)/var_xim)))
+    np.testing.assert_allclose(mean_varxi, var_xip, rtol=0.02 * tol_factor)
+    np.testing.assert_allclose(mean_varxi, var_xim, rtol=0.02 * tol_factor)
+
+
 if __name__ == '__main__':
     test_gg()
     test_spherical()
@@ -2219,3 +2278,4 @@ if __name__ == '__main__':
     test_rperp_local()
     test_oldrperp()
     test_oldrperp_local()
+    test_varxi